backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
fastapi==0.115.12
h11==0.16.0
idna==3.10
pydantic==2.11.3
pydantic[email]
pydantic-settings==2.8.1
pydantic_core==2.33.1
python-dotenv==1.1.0
sniffio==1.3.1
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.1
uvicorn==0.34.0
sqlalchemy
psycopg2-binary
alembic
python-jose[cryptography]
python-multipart
pytest
pytest-asyncio
pytest-xdist
httpx
bcrypt
sendgrid
black
isort
mypy
pytest-cov
watchdog